Move Omnibar code to its own class, and implement OmnibarView
