+++ app/src/main/java/com/duckduckgo/mobile/android/duckduckgo/ui/browser/BrowserFragment.java
@@ -1,7 +1,5 @@
 package com.duckduckgo.mobile.android.duckduckgo.ui.browser;
 
-import android.animation.Animator;
-import android.animation.AnimatorListenerAdapter;
 import android.annotation.SuppressLint;
 import android.app.Activity;
 import android.content.Intent;
@@ -10,17 +8,12 @@ import android.support.annotation.NonNull;
 import android.support.annotation.Nullable;
 import android.support.v4.app.Fragment;
 import android.support.v7.widget.Toolbar;
-import android.view.KeyEvent;
 import android.view.LayoutInflater;
 import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
-import android.view.inputmethod.EditorInfo;
 import android.webkit.WebSettings;
 import android.webkit.WebView;
-import android.widget.EditText;
-import android.widget.ProgressBar;
-import android.widget.TextView;
 
 import com.duckduckgo.mobile.android.duckduckgo.Injector;
 import com.duckduckgo.mobile.android.duckduckgo.R;
@@ -30,9 +23,7 @@ import com.duckduckgo.mobile.android.duckduckgo.ui.browser.web.DDGWebChromeClien
 import com.duckduckgo.mobile.android.duckduckgo.ui.browser.web.DDGWebViewClient;
 import com.duckduckgo.mobile.android.duckduckgo.ui.editbookmark.EditBookmarkDialogFragment;
 import com.duckduckgo.mobile.android.duckduckgo.ui.navigator.Navigator;
-import com.duckduckgo.mobile.android.duckduckgo.util.KeyboardUtils;
-
-import java.lang.ref.WeakReference;
+import com.duckduckgo.mobile.android.duckduckgo.ui.omnibar.Omnibar;
 
 import butterknife.BindView;
 import butterknife.ButterKnife;
@@ -42,7 +33,7 @@ import butterknife.Unbinder;
  * Created by fgei on 5/15/17.
  */
 
-public class BrowserFragment extends Fragment implements BrowserView, OmnibarView {
+public class BrowserFragment extends Fragment implements BrowserView {
 
     public static final String TAG = BrowserFragment.class.getSimpleName();
 
@@ -55,14 +46,8 @@ public class BrowserFragment extends Fragment implements BrowserView, OmnibarVie
     @BindView(R.id.browser_web_view)
     WebView webView;
 
-    @BindView(R.id.appbar_toolbar)
-    Toolbar toolbar;
-
-    @BindView(R.id.appbar_edit_text)
-    EditText searchEditText;
-
-    @BindView(R.id.progress_bar)
-    ProgressBar progressBar;
+    @BindView(R.id.omnibar)
+    Omnibar omnibar;
 
     private Unbinder unbinder;
     private BrowserPresenter browserPresenter;
@@ -79,7 +64,7 @@ public class BrowserFragment extends Fragment implements BrowserView, OmnibarVie
         View rootView = inflater.inflate(R.layout.fragment_browser, container, false);
         unbinder = ButterKnife.bind(this, rootView);
         browserPresenter.attachBrowserView(this);
-        browserPresenter.attachOmnibarView(this);
+        browserPresenter.attachOmnibarView(omnibar);
         return rootView;
     }
 
@@ -164,72 +149,6 @@ public class BrowserFragment extends Fragment implements BrowserView, OmnibarVie
     }
 
     @Override
-    public void displayText(@NonNull String text) {
-        searchEditText.setText(text);
-    }
-
-    @Override
-    public void clearText() {
-        searchEditText.setText("");
-    }
-
-    @Override
-    public void clearFocus() {
-        searchEditText.clearFocus();
-    }
-
-    @Override
-    public void requestFocus() {
-        searchEditText.requestFocus();
-    }
-
-    @Override
-    public void setBackEnabled(boolean enabled) {
-        MenuItem backMenuItem = toolbar.getMenu().findItem(R.id.action_go_back);
-        backMenuItem.setEnabled(enabled);
-    }
-
-    @Override
-    public void setForwardEnabled(boolean enabled) {
-        MenuItem forwardMenuItem = toolbar.getMenu().findItem(R.id.action_go_forward);
-        forwardMenuItem.setEnabled(enabled);
-    }
-
-    @Override
-    public void setRefreshEnabled(boolean enabled) {
-        MenuItem reloadMenuItem = toolbar.getMenu().findItem(R.id.action_refresh);
-        reloadMenuItem.setEnabled(enabled);
-    }
-
-    @Override
-    public void showProgressBar() {
-        if (progressBar.getVisibility() == View.GONE) {
-            progressBar.setVisibility(View.VISIBLE);
-            progressBar.animate().alpha(1);
-        }
-    }
-
-    @Override
-    public void hideProgressBar() {
-        final WeakReference<ProgressBar> progressBarWeakReference = new WeakReference<>(progressBar);
-        progressBar.animate().setStartDelay(400).alpha(0).setListener(new AnimatorListenerAdapter() {
-            @Override
-            public void onAnimationEnd(Animator animation) {
-                super.onAnimationEnd(animation);
-                ProgressBar progressBar = progressBarWeakReference.get();
-                if (progressBar != null) {
-                    progressBar.setVisibility(View.GONE);
-                }
-            }
-        });
-    }
-
-    @Override
-    public void onProgressChanged(int newProgress) {
-        progressBar.setProgress(newProgress);
-    }
-
-    @Override
     public void showConfirmSaveBookmark(@NonNull BookmarkEntity bookmarkEntity) {
         EditBookmarkDialogFragment dialog = EditBookmarkDialogFragment.newInstance(R.string.bookmark_dialog_title_save, bookmarkEntity);
         dialog.show(getFragmentManager(), EditBookmarkDialogFragment.TAG);
@@ -242,8 +161,7 @@ public class BrowserFragment extends Fragment implements BrowserView, OmnibarVie
 
     private void initUI() {
         initWebView(webView);
-        initToolbar(toolbar);
-        initSearchEditText(searchEditText);
+        initOmnibar(omnibar);
     }
 
 
@@ -256,9 +174,9 @@ public class BrowserFragment extends Fragment implements BrowserView, OmnibarVie
         webView.setWebChromeClient(new DDGWebChromeClient(browserPresenter));
     }
 
-    private void initToolbar(Toolbar toolbar) {
-        toolbar.inflateMenu(R.menu.browser);
-        toolbar.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
+    private void initOmnibar(Omnibar omnibar) {
+        omnibar.inflateMenu(R.menu.browser);
+        omnibar.setOnMenuItemClickListener(new Toolbar.OnMenuItemClickListener() {
             @Override
             public boolean onMenuItemClick(MenuItem item) {
                 switch (item.getItemId()) {
@@ -281,26 +199,12 @@ public class BrowserFragment extends Fragment implements BrowserView, OmnibarVie
                 return false;
             }
         });
-    }
-
-    private void initSearchEditText(EditText searchEditText) {
-        searchEditText.setOnEditorActionListener(new TextView.OnEditorActionListener() {
+        omnibar.setOnSearchListener(new Omnibar.OnSearchListener() {
             @Override
-            public boolean onEditorAction(TextView v, int actionId, KeyEvent event) {
-                if (actionId == EditorInfo.IME_ACTION_SEARCH || wasEnterPressed(event)) {
-                    String text = v.getText().toString().trim();
-                    browserPresenter.requestSearch(text);
-                    KeyboardUtils.hideKeyboard(v);
-                    v.clearFocus();
-                    return true;
-                }
-                return false;
+            public void onTextSearched(@NonNull String text) {
+                browserPresenter.requestSearch(text);
             }
         });
-    }
-
-    private boolean wasEnterPressed(KeyEvent event) {
-        return event.getAction() == KeyEvent.ACTION_DOWN && event.getKeyCode() == KeyEvent.KEYCODE_ENTER;
     }
 
     private void handleBookmarkResult(int resultCode, Intent data) {
